More async handling porting work
This patch completes the first pass of conversions from a single newest
entry index structure to the 2 member array which holds sync'd and
unsync'd "newest-entries".

Many raft_instance struct parameters are no longer 'const' since the
functions need to take the ri_mutex in order to access the
newest-entries data.