Add raft server sync thread
When the raft service is started w/ asynchronous write support a thread
will be spawned which will issue raft_server_backend_sync_pending()
after sleeping for some time.  The thread may also have to issue an RPC
to the leader so the leader may update its next and commit index values.

This is still a draft.