raft_server_process_append_entries_term_check_ops() fixes
The main fix is for the below stack trace.  The problem was caused when
the leader was paused in gdb and resumed some seconds later.  A new
leader had sent this server an append-entry request which caused the
server to enter raft_server_append_entry_check_already_stored() and then
assert on a check for RAFT_STATE_FOLLOWER.  The issue is that at this
point, the server has accepted a follower role but did not change its
ri_state value to record this fact.

raft_server_process_append_entries_term_check_ops() has been changed to
cause a stale leader to call raft_server_becomes_follower().

The other fix is for RAFT_STATE_CANDIDATE.  The code was only calling
raft_server_becomes_follower() when the remote term value was == to this
server's term value.  This check should be widened so that higher term
values will also cause the server to become a follower.

gdb) bt
    ri=0x6390c0 <raftInstance>, raerq=0x63b328 <sink_buf+40>)
    at src/raft_server.c:2154
    ri=0x6390c0 <raftInstance>, raerq=0x63b328 <sink_buf+40>,
    reh=0x7fffffffb2b0) at src/raft_server.c:2300
    ri=0x6390c0 <raftInstance>, raerq=0x63b328 <sink_buf+40>)
    at src/raft_server.c:2337
    ri=0x6390c0 <raftInstance>, sender_csn=0x612000001240,
    rrm=0x63b300 <sink_buf>) at src/raft_server.c:2587
    ri=0x6390c0 <raftInstance>, rrm=0x63b300 <sink_buf>,
    sender_csn=0x612000001240) at src/raft_server.c:2878
    ri=0x6390c0 <raftInstance>, recv_buffer=0x63b300 <sink_buf> "\003",
    recv_bytes=104, from=0x63b2c0 <from>) at src/raft_server.c:2932
    at src/raft_net.c:1043
    epm=0x63a7a8 <raftInstance+5864>, timeout=-1) at src/epoll_mgr.c:146
    at src/raft_server.c:4004
    raft_uuid_str=0x7fffffffd3a2 "ebca3cf0-f2c4-11ea-ac72-90324b2d1e89",
    this_peer_uuid_str=0x7fffffffd3ca "ebcba676-f2c4-11ea-8275-90324b2d1e89",
    sm_request_handler=0x40f85b <raft_server_test_rst_sm_handler>,
    type=RAFT_INSTANCE_STORE_POSIX_FLAT_FILE, sync_writes=false, arg=0x0)
    at src/raft_server.c:4043
    at test/raft_server_test.c:623