RAFT_RPC_MSG_TYPE_SYNC_IDX_UPDATE rpc msg type
RAFT_RPC_MSG_TYPE_SYNC_IDX_UPDATE has been added so that the raft leader
may be immediately notified of followers' sync completions.  The goal is
to reduce commit latency by having the leader learn the followers'
sync-indexes without having to wait for a reply of a future AE request.

The downside of this approach is that the followers will send extra msgs
to the leader which may increase latency for other operations.  I think
it may be possible to filter some of these messages by tracking the
last-sync-idx sent to the leader and bypassing sends when that value has
not changed.