Add context awareness to epoll handle destruction
This patch introduces a feature which allows for async removal of epoll
handles.  The async behavior occurs when the user has supplied a ref_cb
and the thread calling epoll_handle_del() is different from the thread
that blocks in epoll_wait().  The ref_cb allows the add and del methods of
epoll_mgr to take and release a ref.  This functionality is basically a
copy of Kit's getput_cb which is introduced in the tcp branch.

At the moment, nothing uses the async removal. This will have to be
rectified in later commits on this branch which implement unit tests for
this feature.

Note that this patch will now enforce thread affinity with an epoll_mgr
instance.  Once a thread has entered
epoll_mgr_wait_and_process_events(), it will affine itself to the
epoll_mgr struct and later calls to epoll_mgr_wait_and_process_events()
will assert that the thread id has not changed.