PumiceDB client fix of 'pmdb-write-pending'
'pmdb-write-pending' is a ctl-interface output found at
/pumice_db_test_client/pmdb-test-apps.  The server was comparing two
term values to determine if this should be set to true in the RPC reply.
However, both values being compared were set to -1, meaning that both
the term values were purposely marked as invalid.  'pmdb-write-pending'
should have been marked 'false' in this case.