raft_leader_instance_is_fresh() bug fix
raft_leader_instance_is_fresh() was returning 'false' due to an
incorrect underflow check.  The check was ensuring a negative value was
not considered, however, it was also causing a value of '0' to be
bypassed.  This is incorrect due to the coarse granularity of the clock
source which is 1ms.