Raft leader issues AE before sync'ing its own entry
This patch allows the raft leader to send out append entry requests to
followers which is has not yet synced.

The append entry send routine may be touched off by the sync thread
should it believe that the last sync would allow for the commit idx to
be incremented.

The pmdb server now has an option to run the server in async mode.