Allow for commit idx to decrease while establishing leadership
Async entry append has revealed a defect in the method used by the
leader for calculating the commit index.  The issue is that after a
leadership change, the node set comprising the most recent quorum (ie
the quorum which 'approved' the latest commit) may have some members -
other than the newly appointed leader - die or become latent.  When the
new leader's remaining followers having lagging indexes, a commit index
calculation done on the leader will yield a value which is < the
leader's current commit idx value.  Prior to this patch such a situation
would cause an abort since the leader would panic if it detected its
followers were reverting their commit idx values.  However, this is too
strict.  raft_server_leader_can_advance_commit_idx() performs a check
which detects if the current quorum has yet committed the leader's
initial term idx (this is the leader change marker commit which every
new leader performs at the start of its term).  This check is now used
to limit the scope of the assert to allow for a temporary regression in
the commit idx value until after all alive peers have applied the
leader-change-marker commit.