Issue #27: Raft leader considers follower's sync-idx for next-idx
The method for determining a peer's next-idx was taken very literally
from the raft paper.  Recent testing showed that this brute force
discovery method was highly inefficient.  Its now replaced with a rather
simple approach which uses the existing sync-idx in the AE reply RPC.
The follower's sync-idx may still need to be rolled back, however, it
is still a good starting place for the leader's discovery of that
follower's next-idx.