Issue #28: Newly started peer applies interim commit-idx value
During scale testing it was noticed that, on reboot, a lagging peer
would not increment the commit index.  Previously, the code would not
consider advancing the commit index if
raft_server_append_entry_log_prepare_and_check() returned any error.
However, there are 2 known error cases, -EALREADY and -ERANGE, where
commit idx advancing may still occur so long as the idx value is the
lesser of sync-idx or leader-commit-idx.

Without this patch, a lagged follower could delay large huge numbers of SM
apply operations until it had fully caught up with the leader.  The
result is that a huge batch of SM apply operations (ie advancing
last-applied-idx) could take almost as long as the recovery itself.
Note that it is detrimentral for a raft system to have large skews
between commit-idx and last-applied-idx.  In systems with just enough
nodes to form a quorum, it may be the case that the leader blocks client
operations until the lagged followers (or itself) has applied that
leader's first transaction.