Modify raft_net to have separate client and server tcp_mgrs
In order to have greater configuration flexibility, separate the unitary
tcp_mgr in raft_net into two, one for clients and one for peers.

This requires moving the connection setup to the connect phase, as connection
type cannot be determined during raft boot. A connect_info callback has
also been added to allow raft_net to tell tcp_mgr the ipaddr and port of
the remote node when needed.