Generate checkpoints in raft-server
Patch implements callback in src/raft_server_backend_rocksdb.c for
rib_backend_checkpoint() and adds a new thread in raft_server which will
launch if the callback has been assigned.

RAFT_LREG_CHKPT_IDX is now writable through the ctl-interface.  The
supplied value is ignored and the only effect is that ri_take_checkpoint
is set to true which causes the checkpoint thread to take a checkpoint
at its next wake up (currently a 1 second interval).

rsbr_checkpoint() builds a subdir per checkpoint which is based on the
db-uuid, peer-uuid, and sync-idx at checkpoint time.  The checkpoint is
made in a tmp dir (prefixed with ".in-progress_").  On successful
completion of the checkpoint, the directory is renamed.

rib_backend_checkpoint() has been changed to return the value of the
sync_idx at the time just prior to the checkpoint creation.  This is so
the ri_checkpoint_last_idx value can be updated in raft_server.c.