Implement raft log reaping
This patch starts to handle ri_lowest_idx.

Reaping of older log entries is now enabled as is the scanning process
(at startup) to locate the lowest raft entry.

raft_server_entries_scan() now scans the beginning of the raft entry
keyspace as a sanity check since that portion could be bypassed due to
raftPersistentAppMaxScanEntries.