raft_server_rocksdb_release_cf_table() called in scope
Prior to this patch, the cf_table was being released from inside
rsbr_destroy().  This prevented reused of the table within the same
instance of raft_server_instance_run().

This patch causes the release of the cf table to occur in the same scope
as where it was initialized / allocated.