Raft shutdown error checking is more stringent
The recovery code is going to shutdown the raft subsystems and restart
them after the external bulk recovery process has completed.  In order
for the process to safely restart itself in-process, we should be very
certain that the shutdown procedures occurred as expected so subsequent
restarts do not encounter leftover state.

In functions where there are multiple cleanup functions, the policy is
to return the first non-zero error code encountered but to log all
errors.