Replace raft instance simple counters for evp and eph
The counters being used to track the numbers of evp and eph in use were
not intended for use where the raft instance structure would be
'restarted'.  Basically there was no release() method for either and
even worse, some users would index directly into the ri_evps using an
enum as a indexer.

This patch maintains the arrays for both but uses a looping lookup
method to find / allocate / release the structure based on the 'type'
(enum raft_event_pipe_types) or the fd number.

The patch also refines some logging so that startup errors do not result
in a cascade of error msgs in shutdown.  ri_startup_error has been added
to designate a startup failued.  The shutdown will still attempt to run
but using a lower log level.