Add LREG_INIT_OPT_INLINED_MEMBER feature to registry
LREG_INIT_OPT_INLINED_MEMBER allows a child installed with this
option to be added to parent which itself has not yet been installed.
LREG_INIT_OPT_INLINED_MEMBER implies that the child is a structural
subcomponent of the parent, such as in the case where a struct with an
lreg has an array of histograms where each has an lreg.

LREG_INIT_OPT_INLINED_MEMBER will simplify add and remove operations
since the registry thread will only see a single request.