Async install additions for registry
registry.c wasn't providing a callback for reg nodes which were queued
but not yet installed.  In this case, a pointer is being held on the
install queue which may need to be unref'd for garbage collection.  To
help with this a new cb op has been added,
LREG_NODE_CB_OP_INSTALL_QUEUED_NODE, which is called from installer's
thread context.  Most of the patch deals with adding the
LREG_NODE_CB_OP_INSTALL_QUEUED_NODE case to the many lreg cb functions.