Issue#109 Corrected data type for var in pmdb_sm_handler_client_read
In pmdb_sm_handler_client_read() unsigned size_t was used
to store return value of pmdb_object_lookup().
But the return value was checked for negative error code.
Which was making recipe fail as it expects the status to
be ENOENT and E2BIG was return.
Changed the data type to ssize_t.