lreg_node_remove() implementation and unit test
lreg_node_remove() is now present and can remove lreg nodes either
synchronously or async depending on the thread context.

A new union member, lrn_removal_lentry, has been added so that it does
not conflict with the stailq entry needed for the removal queue.

Added an enum lreg_node_states which now holds the LREG node
installation states.

New install state has been added for removing and some inline functions
to transfer the state of the lreg node from installed to removing to
uninstalled.

Renamed the macro generated lreg nodes so that a double underscore
precedes the supplied name.