Fix log.h lreg handling
struct log_entry_info now initializes its lrn_head using
CIRCLEQ_HEAD_INITIALIZER().

lrn_inlined_member and lrn_inlined_children are used on child and
parent log nodes so that we are not required to wait for async
initialization to complete before applying a child into a parent.  Since
all of these objects are statically defined this should be safe.