Recovery peer implementation progress
The code will now enter rsbr_bulk_recover_import_remote_db() and build
the local and remote db paths.  The rsync is not yet in place..

This patch removes rsbr_bulk_recover_prepare() since it was not usable
because the db is not open when it is called.  Instead, we have added
raft_server_chkpt_prior_to_recovery(), which takes a checkpoint prior to
closing the db inside raft_server_instance_run().