Add popen_cmd.c
popen_cmd provides a simple interface to popen(3).  I decided to go this
route instead of directly using fork + exec for launching externals
apps.  Once nice advantage of popen() is that its use of the FILE
interface helps with newline delineation.

The rest of the patch is a small unit test and addition to the Makefile.