raft_bulk_recovery: implement checkpoint dir scanning
This patch enables the server to determine its last-checkpoint-idx based
on the contents of the chkpt/self/ directory.  The scan detects stale
UUID's and causes those checkpoints to be placed into the trash.

The scan is called during bootup and following a checkpoint to remove
checkpoint(s) which are no longer needed as determined by the
ri_num_checkpoints value.  The scan function, rsbr_self_chkpt_scan(),
uses scandirat() which returns the list of entries in sorted order from
lowest checkpoint-idx to highest.