raft_bulk_recovery: implement rsbr_bulk_recovery_db_scrub()
rsbr_bulk_recovery_db_scrub() will scan and modify the raft entry
headers contained in the recovered checkpoint and place the uuid of this
peer into each header.

This patch adds a new KV which is used to stash the uuid of the
recovered DB.  The uuid, stored in ri_db_recovery_uuid, is used by
rsbr_bulk_recovery_db_scrub() to resume a "mostly complete" previous
recovery attempt which was cut short somewhere in between completing the
scrub and renaming the "in progress" recovery db dir.  In this case, the
recovery will resume on the next instance of the raft process.  Note
that the scrubbing procedure is always started from scratch following an
interruption.

Note this patch is still a WIP.