raft_bulk_recovery: fixes and promotion of scrubbed db
This patch adds the routine, rsbr_bulk_recovery_promote_scrubbed_db(),
which renames the "scrubbed" db (which is the recovery source) from its
temporary name to the default name "db/".

Remove stale comments and unused code.

Fixes
The previous patch was still a WIP and some minor fixes were required to
get the recovery operational.

rsbr_bulk_recover_build_remote_path() - thinko where reh.reh_index was
not being set.

rsbr_entry_header_write_recovery_scrub() - sync writes were way too slow
so we're using async.  This is ok since the log header write, which is
synchronous so it flushes the rocksdb WAL, is performed after the entry
header rewrites.  Note that if the entry header rewrite process is
"resumable" following a crash.

rsbr_init_header() - fix a thinko where memset() was occurring in the
case of raft_instance_is_recovering().  This is wrong since we need to
preserve the contents of the previous db instance.