raft_bulk_recovery: add raft_server_entry_read_by_store_common()
raft_server_entry_read_by_store_common() is now used for reading raft
entries and entry-headers.  The function contains a range check which
catches the situation where the checkpoint thread has reaped the
lowest-idx while a read operation was in progress.  The approach is to
test the ri_lowest_idx value against the requested index when the
backend returns -ENOENT.  This method is an alternative serializing read
and checkpoint/reap operations since the latter may take considerable
time.

The patch also swaps the order of operations in raft_server_reap_log()
so that the ri_lowest_idx value is set before the reap operation is
enacted.