raft_bulk_recovery: Issue #28 regression fix
raft_server_process_append_entries_request() had been changed to accept
EALREADY and ERANGE as errors which would allow the commit index to be
updated.  However, ERANGE is not eligible for allowing this behavior
since it may, and likely does, indicate that the rollback to an entry in
common with the leader is incomplete.  Allowing for commit and SM apply
to happen in this context is at best wrong and at worst a source of DB
corruption.