raft_ackd_idx: Add rfi_ackd_idx and factor into commit-idx calculation
This patch qualifies the sync-idx value, used by the leader to
determine the cluster's sync-idx, as being acknowledged during the
leader's session.  In other words, the sync-idx will only be considered
once the leader and follower have converged on the lowest entry with a
common term value.  This will prevent recovered followers from having
unqualified indices considered in the commit index calculation so that
the cluster avoids committing and applying stale entries on one or more
followers.