automake: Makefile.am directly links registry-test w/ core .o files
registry-test must access init_ctx() of the niova library (see init.c).
To do this, we link registry-test directly with the niova-core .o
objects so that there's .init section for everything.  Note that when
using the shared library, libniova.so, the library's .init section will
be executed before the linking binary's.