io.c: io_iovs_advance() and io_iov_restore()
These functions are iov helpers for supporting the continuation of
incomplete vectored operations, such as readv() and writev().

The pair of functions enables for the non-destructive mgmt of iovs w/out
requiring copies of the entire array.  This is found to be helpful when
using io_uring to process an iov array through a pipeline containing
more than one file descriptor.