Issue#58 PmdbObjGetX should return actual reply size.
Added parameter to PmdbObjGetX() to get the actual size
of the value returned from server.
This parameter is important for application to
retry the read request in case buffer allocated to
read the value is insufficient. Looking at the
reply_size, application can allocate bigger buffer
and retry the operation.