Methodize pmdb go server library calls
WriteKV(), ReadKV(), CopyDataToBuffer(), etc. are now presented as
methods, as is PmdbStartServer() whose method is called 'Run()'.

I added an interface, PmdbServerAPI, to replace the function callback
types.  This interface can be implemented by the user's server-side
implementation.  In this patch, the dictionary app has been given a
object, DictionaryServer, which implements both 'Apply()' and 'Read()'.
Therefore, an instance of DictionaryServer may fulfill the requirements
of being an implementation of PmdbServerAPI.