Issue85 Bug fix for accessing read value buffer out of bound.
rocksdb_get_cf() returns the value of the key as char * buffer.
Also stores the length of the value in the value_length pointer.

During tests with lkvt, it failed with:
slice bounds out of range [:33] with length 30

That means buffer length is 30 bytes and pmdb server go library
is trying to access till offset 33.

Replaced GoString() with GoStringN() to convert the C char* string
to golang string.