Fix pmdb client request timeout bug
pmdb_client_error_demonstration2.yml was failing due to the tv_nsec
member of the timeout not being initialized.  The source of this value
was coming from the stack in pmdbtc_submit_request().

This patch initializes the stack var as well as the tv_nsec member
inside of pmdb_request_options_init().