pmdb_sm_handler_pmdb_sm_apply() derives cowr_sa from rncr
Previously, pmdb_sm_handler_pmdb_sm_apply() was performing a lookup into
the co RB_TREE to obtain the pmdb_cowr_sub_app pointer associated with
the write request.  This caused an additional lookup on an already
referenced pointer as well as an additional 'put'.

This patch stashes the pointer inside the rncr and uses that pointer in
the apply handler.  In addition, the patch performs strict checks on the
pmdb_cowr_sub_app to ensure that it matches the rncui context of the
request.