Fix problem introduced in b9e45a
Manisha's review of the code had found that the rncr was being used in
an incorrect manner.  The patch was improperly assuming that the rncr
contents from the 'write' context would be presented in 'apply'
context.  This patch reverts parts of b9e45a while keeping the stricted
checks in place.