Simplify go task and waitgroup usage
The previous code was using a "closure" to perform sendReq().  The patch
removes the closure and places the waitgroup Add() / Done() directly
within sendReq().

This patch also appears to fix a bug in the former closure usage where
`index int` was being used as an array indice without initialization.

Lastly, the patch uses getServerAddr(false) within the loop to prevent
superfluous connections to the gossip layer.