Write request may get lost due while flushing coalesced buffer.
The coalesced buffer is sufficient to accomodate
the incoming write request it checked even after
including that request into the write supplement.
Because of that, write request will be present in
the write supplement but its corresponding entry
in raft header will be missing.

We should flush the buffer even before write
request gets added to the write supplement.