Bulk recovery fix: update_sli
update_sli stands for "update synced log index".  It's used for recently
bulk-recovered to notify the leader to reset its raft_follower_info
state.

Note that this change should be accompanied by a check of the raft db
uud, which is included in all raft rpc msgs.  This would enhance the
check allowing for update_sli to be applied only in case where the
leader has detected a db-uuid change (which implies the follower
underwent bulk-recovery).

Another change here is in raft_server_refresh_follower_prev_log_term()
which removes code that increased rfi_next_idx if it fell below the
leader's lowest_idx.  This was put in place to prevent the leader from
attempting reads of compacted entries.  The problem with this method was
that it prevented the -ERANGE from being detected in
raft_server_read_entry_register_idx().  This error is used to notify the
follower to enter bulk recovery.  Note that
raft_server_read_entry_register_idx() exists to prevent such reads and
the increasing of next_idx.

raft_server_leader_init_append_entry_msg() fix sets
raerqm_prev_log_index in the case where the ERANGE error is seen.