Fix pmdb_cowr_sub_app_tree bug caused by leader demotion
The pumicedb server utilitizes an REF_TREE when coalescing is enabled
so that it can identify and filter retried requests which may not be
identified by the typical raft and pumicedb protocols.  The
pmdb_cowr_sub_app type is the object which represents a pending request
residing in the coalesce buffer.  These objects are strictly checked in
pmdb_sm_apply for correctness against the coalesce buffer contents.

The bug occurred when a leader was demoted with pending items in its
pmdb_cowr_sub_app_tree.  Later, when this server would apply a commit as
a follower the strict checks referenced above would fail.

This patch adds a new function, pmdb_cowr_sub_app_release_all(), which
is called when a server is not in leader state.  This function causes
all tree items to be removed.