Functionality to create rocksdb snapshot.
For range read to work across iterations, rocksdb
snapshots will be created and maintined in an
RB tree. If range read requests needs to be span
acorss multiple iterations, we can use same snapshot
and avoid reading new writes which can happen in between.

The Snapshot tree will be maintained at leader and will
be destroyed on leadership change.

The snapshot items will be maintained in an LRU list as well.
So that snapshots which are created for really long time and
didn't get destroyed for any reason, can be released.