Address shortcomings in lreg handling / ownership
Transferring ownership of the registry operations had several problems.
First, the prior code only considered the ctl-interface event source
(inotify) when unregistering from the util-thread.  This left the
registry fd behind and allowed for the util-thread to still accept
lreg_node installations - which violating basic coherency rules since
there's no locking in the registry.

lreg_remove_event_src() is introduced and added to
lctli_util_thread_unregister() to address the above.

When addding and removing an lreg_node, there are situations where the
user needs to know if these operations have completed.
lreg_node_wait_for_install_state() has been added to serve this purpose.
Note that the usage semantics are somewhat complicated by the fact that
the lreg_node can advance 'past' a single requested state.  For
instance, if a user wants to know if the node has been installed, it can
request to wait for LREG_NODE_INSTALLED, however, it may be the case
that the node was quickly installed and then removed.  To deal with
this, the lreg_node_states have been made into a mask so that users of
lreg_node_wait_for_install_state() can supply multiple states.

Event handling is now down by an eventfd rather than an ev-pipe.
lreg_notify() is used for notifying the registry layer that an lreg_node
has been queued.