pmdb_cowr_sub_app_add() uses msg_id to detect retries
PMDB must enforce a strict policy of 'queue_depth == 1'.  The modified
code deals with the detection of multiple writes to the same rncui
within the scope of the current set of pending writes (in the coalesce
buffer).  This patch changes the comparison item from the client_uuid to
the msg_id since the client_uuid may for a proxy (which itself is
representing a set of clients).