Remove tmc_event_recvd_while_handoff
The original idea behind this flag was to allow for the tmc to be
immediately requeued if additional sends had arrived on the 'fd' while
it was being processed (or queued waiting for processing).  However,
it's not entirely obvious how to do this, especially when using just a
single worker thread, without potentially starving other connections.
Instead, we'll rely on epoll to immediately post an event if unprocessed
data remains on the fd.