Add write mutex to raft server / split out client recv handlers
The client recv handler functions have now been moved to their own
routines, out of raft_server_client_recv_handler().  Additionally, the
rncr gets prepare() and complete() routines which are responsible for
setting up the rncr and verifying the request state prior to proceeding.

The patch also places the write mutex at several places in an attempt to
fully serialize the various components of the write pipeline.  This is
so the client recv handler may be run from one or more threads.

The serialization points are:
* raft_server_leader_co_wr_timer_expired() - here the coalesce buffer is
  tested and flushed.
* raft_server_do_client_write() - (formerly
  raft_server_client_recv_handler()) wraps the calling of the state
  machine callback all the way through
  raft_server_client_rncr_write_raft_entry(), which is conditionally
  called from raft_server_client_rncr_complete().
* raft_server_sm_apply_evp_cb() - which wraps
  raft_server_state_machine_apply().

Note that I'm not fully confident that this concurrency effort is bug
free.  However, this approach should meet the primary requirement of
removing complex read operations from the main raft thread.

I'm now of the opinion that the only way to fully address the
potential issues and requirements of raft is to use io_uring along with
the nconn.c interface currently residing in niova-block.  This would
handle concurrent, async read and write replies to a single client
without having to explicitly serialize at the socket (as is being done
now per this branch).