write_prep callback can return if write needs to continue.
For the refresh lease operation, lease will be refreshed
on leader only and will not continue with raft write replication.
To indicate pumice layer not to continue write even if refresh
operation was succesful, continue_wr flag was introduced.
Application layer can return success and set continue_wr to 0
to indicate not to continue the raft write operation.

The return code to the client layer will indicate not to retry
the operation and consider the operation as success.