Progress on buffer set cache implementation / Use array over tree
This patch changes the previous where an RB tree was used to hold the
items.  Instead, we simply use a logical 2D array which is sized
according to the next prime value to follow the number of items in the
set.  Since we're not intending to utilize ref counts for the cache
items an array seems more reasonable.

The code here deals w/ hash bucket collisions and should ensure the
LRU'ness of the items.

to-be-squashed-1680375695