Add buffer-set api call for user caching
buffer_set_user_cache_release_item() is introduced here.  It allows a
buffer-set user to maintain a pointer to a 'freed' item.  The idea here
is that a user may keep a pointer to the item as long it provides a
revoke callback to buffer_set_user_cache_release_item().  Note that
buffer_set_release_item() may be called after
buffer_set_user_cache_release_item() to fully release the item - in this
case, the callback will not be issued.