Fix niova_io_iovs_advance() bug when iovs are completely used
Calling niova_io_iovs_advance() with parameters that completely 'use'
the entire iov set caused the function to assert on an internal check
that verified the ensuing index was < the number of iovs.

The approach taken here is to detect the condition and return the error,
-EXFULL, rather than modify the last iov such that its base pointer
would be out of bounds and its length would be 0.